LRQA-40276 Make AutoCloseRule an inner class
