+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/AutoCloseUtil.java
@@ -16,9 +16,12 @@ package com.liferay.jenkins.results.parser;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.tools.ant.Project;
@@ -416,6 +419,157 @@ public class AutoCloseUtil {
 		}
 
 		return false;
+	}
+
+	private static class AutoCloseRule {
+
+		public AutoCloseRule(String ruleData) {
+			this.ruleData = ruleData;
+
+			String[] ruleDataArray = ruleData.split("\\|");
+
+			rulePattern = Pattern.compile(ruleDataArray[0]);
+
+			if (ruleDataArray[1].endsWith("%")) {
+				String percentageRule = ruleDataArray[1];
+
+				maxFailPercentage = Integer.parseInt(
+					percentageRule.substring(
+						0, percentageRule.length() - 1)) / 100;
+			}
+			else {
+				maxFailCount = Integer.parseInt(ruleDataArray[1]);
+			}
+		}
+
+		public List<Build> evaluate(List<Build> downstreamBuilds) {
+			downstreamBuilds = getMatchingBuilds(downstreamBuilds);
+
+			List<Build> failingInUpstreamJobDownstreamBuilds = new ArrayList<>(
+				downstreamBuilds.size());
+
+			for (Build downstreamBuild : downstreamBuilds) {
+				if (UpstreamFailureUtil.isBuildFailingInUpstreamJob(
+						downstreamBuild)) {
+
+					failingInUpstreamJobDownstreamBuilds.add(downstreamBuild);
+
+					continue;
+				}
+
+				List<TestResult> testResults = new ArrayList<>();
+
+				testResults.addAll(downstreamBuild.getTestResults("FAILED"));
+				testResults.addAll(
+					downstreamBuild.getTestResults("REGRESSION"));
+
+				boolean containsUniqueTestFailure = false;
+
+				if (testResults.isEmpty()) {
+					containsUniqueTestFailure = true;
+				}
+				else {
+					for (TestResult testResult : testResults) {
+						if (!UpstreamFailureUtil.isTestFailingInUpstreamJob(
+								testResult)) {
+
+							containsUniqueTestFailure = true;
+
+							break;
+						}
+					}
+				}
+
+				if (!containsUniqueTestFailure) {
+					failingInUpstreamJobDownstreamBuilds.add(downstreamBuild);
+				}
+			}
+
+			downstreamBuilds.removeAll(failingInUpstreamJobDownstreamBuilds);
+
+			if (downstreamBuilds.isEmpty()) {
+				return Collections.emptyList();
+			}
+
+			List<Build> failedDownstreamBuilds = new ArrayList<>(
+				downstreamBuilds.size());
+
+			int failLimit = 0;
+
+			if (maxFailPercentage != -1) {
+				failLimit = (int)(maxFailPercentage * downstreamBuilds.size());
+
+				if (failLimit > 0) {
+					failLimit--;
+				}
+			}
+			else {
+				failLimit = maxFailCount;
+			}
+
+			for (Build downstreamBuild : downstreamBuilds) {
+				String status = downstreamBuild.getStatus();
+
+				if (!status.equals("completed")) {
+					continue;
+				}
+
+				String result = downstreamBuild.getResult();
+
+				if ((result != null) && !result.equals("SUCCESS")) {
+					failedDownstreamBuilds.add(downstreamBuild);
+				}
+			}
+
+			if (failedDownstreamBuilds.size() > failLimit) {
+				return failedDownstreamBuilds;
+			}
+
+			return Collections.emptyList();
+		}
+
+		@Override
+		public String toString() {
+			return ruleData;
+		}
+
+		protected String getBatchName(Build build) {
+			String batchName = build.getParameterValue("JOB_VARIANT");
+
+			if ((batchName == null) || batchName.isEmpty()) {
+				batchName = build.getParameterValue("JENKINS_JOB_VARIANT");
+			}
+
+			return batchName;
+		}
+
+		protected List<Build> getMatchingBuilds(List<Build> downstreamBuilds) {
+			List<Build> filteredDownstreamBuilds = new ArrayList<>(
+				downstreamBuilds.size());
+
+			for (Build downstreamBuild : downstreamBuilds) {
+				String batchName = getBatchName(downstreamBuild);
+
+				if ((batchName == null) || batchName.isEmpty()) {
+					continue;
+				}
+
+				Matcher matcher = rulePattern.matcher(
+					getBatchName(downstreamBuild));
+
+				if (matcher.matches()) {
+					filteredDownstreamBuilds.add(downstreamBuild);
+				}
+			}
+
+			return filteredDownstreamBuilds;
+		}
+
+		protected int maxFailCount = -1;
+		protected float maxFailPercentage = -1;
+		protected String ruleData;
+		protected Pattern rulePattern;
+
 	}
 
 }
